Testing support rails under 5.1.0beta1
