+++ lib/sqlogger/railtie.rb
@@ -4,8 +4,8 @@ module Sqlogger
   class Railtie < Rails::Railtie
     initializer "sqlogger_railtie.configure_rails_initialization" do
       ActiveSupport.on_load :active_record do
-        require 'sqlogger/monkey/active_record/replace_log_subscriber'
-        ActiveRecord::LogSubscriber.send :include, Sqlogger::Monkey::ActiveRecord::ReplaceLogSubscriber
+        require 'sqlogger/monkey/include_active_record_log_subscriber'
+        ActiveRecord::LogSubscriber.send :include, Sqlogger::Monkey::IncludeActiveRecordLogSubscriber
       end
     end
